+++ src/cqspoolterrain/__init__.py
@@ -1,2 +1,3 @@
 from .Base import Base
 from .Spool import Spool
+from .Cradle import Cradle