+++ changes.md
@@ -2,3 +2,4 @@
 
 ## Main wip
 * Added Initial Spool Mockup.
+* Added Initial Cradle and example